[Jeet] Refactored the code to reduce n numbers of line in actual module file
