+++ plugins/module_utils/manage_resource_manager/nd_manage_resource_manager_resources.py
@@ -4,9 +4,7 @@
 from __future__ import annotations
 
 import copy
-import ipaddress
 import logging
-from typing import Any
 
 from ansible_collections.cisco.nd.plugins.module_utils.common.pydantic_compat import ValidationError
 from ansible_collections.cisco.nd.plugins.module_utils.nd_v2 import NDModule
@@ -25,12 +23,6 @@ from ansible_collections.cisco.nd.plugins.module_utils.models.manage_resource_ma
 )
 from ansible_collections.cisco.nd.plugins.module_utils.models.manage_resource_manager.resource_manager_request_model import (
     ResourceManagerBatchRequest,
-    ResourceManagerRequest,
-    FabricScope,
-    DeviceScope,
-    DeviceInterfaceScope,
-    DevicePairScope,
-    LinkScope,
 )
 from ansible_collections.cisco.nd.plugins.module_utils.models.manage_resource_manager.resource_manager_response_model import (
     ResourcesManagerBatchResponse,
@@ -41,794 +33,19 @@ from ansible_collections.cisco.nd.plugins.module_utils.endpoints.v1.manage.manag
     EpManageFabricResourcesActionsRemovePost,
 )
 from ansible_collections.cisco.nd.plugins.module_utils.common.exceptions import NDModuleError
-from ansible_collections.cisco.nd.plugins.module_utils.models.manage_resource_manager.constants import (
-    API_SCOPE_TYPE_TO_PLAYBOOK,
-    POOL_SCOPE_MAP,
-)
 from ansible_collections.cisco.nd.plugins.module_utils.models.manage_switches.switch_data_models import (
     SwitchDataModel,
 )
 from ansible_collections.cisco.nd.plugins.module_utils.fabric_inventory import FabricSwitchInventory
+from ansible_collections.cisco.nd.plugins.module_utils.manage_resource_manager.resource_manager_diff import ResourceManagerDiffEngine
+from ansible_collections.cisco.nd.plugins.module_utils.manage_resource_manager.resource_manager_helpers import ResourceManagerResourceHelpersMixin
 
 # =========================================================================
-# Validation & Diff
+# Resource Manager module
 # =========================================================================
 
 
-class ResourceManagerDiffEngine:
-    """Provide stateless validation and diff computation helpers."""
-
-    @staticmethod
-    def _normalize_pool_name(pool_name: str, log: logging.Logger) -> str | None:
-        """Normalize pool_name to canonical constant form based on ``POOL_SCOPE_MAP`` keys.
-
-        Converts API-style names like ``loopbackId`` to playbook constant names like
-        ``LOOPBACK_ID`` when a token-equivalent key exists in ``POOL_SCOPE_MAP``.
-
-        Args:
-            pool_name: Raw pool name from config or API.
-            log: Logger instance.
-
-        Returns:
-            Canonical pool constant when recognized, otherwise the stripped input value.
-        """
-        if pool_name is None:
-            log.debug("_normalize_pool_name: pool_name is None, returning None")
-            return None
-
-        raw = str(pool_name).strip()
-        if not raw:
-            log.debug("_normalize_pool_name: pool_name stripped to empty string, returning ''")
-            return raw
-
-        token = "".join(ch.lower() for ch in raw if ch.isalnum())
-        if not token:
-            log.debug(
-                "_normalize_pool_name: no alphanumeric chars in pool_name='%s', returning raw='%s'",
-                pool_name,
-                raw,
-            )
-            return raw
-
-        canonical_by_token = {"".join(ch.lower() for ch in key if ch.isalnum()): key for key in POOL_SCOPE_MAP}
-        result = canonical_by_token.get(token, raw)
-        if result != raw:
-            log.debug(
-                "_normalize_pool_name: pool_name='%s' normalized to canonical='%s' (token='%s')",
-                pool_name,
-                result,
-                token,
-            )
-        else:
-            log.debug(
-                "_normalize_pool_name: pool_name='%s' not found in POOL_SCOPE_MAP by token='%s', returning raw='%s'",
-                pool_name,
-                token,
-                raw,
-            )
-        log.debug("Returning normalized pool_name='%s' from raw='%s' ", result, raw)
-        return result
-
-    @staticmethod
-    def _normalize_entity_key(entity_name: str, log: logging.Logger) -> str:
-        """Normalize entity_name for order-insensitive comparison.
-
-        Args:
-            entity_name: Raw entity name string.
-            log: Logger instance.
-
-        Returns:
-            Tilde-separated string with parts sorted alphabetically.
-        """
-        normalize_entity_name = "~".join(sorted(entity_name.split("~")))
-        log.debug("Returning normalized entity_name='%s' from raw='%s'", normalize_entity_name, entity_name)
-        return normalize_entity_name
-
-    @staticmethod
-    def _resource_attr(resource, model_attr: str, dict_key: str | None = None):
-        """Return a field from either a ResourceManagerResponse model or raw API dict."""
-        if hasattr(resource, model_attr):
-            return getattr(resource, model_attr)
-        if isinstance(resource, dict):
-            return resource.get(dict_key or model_attr)
-        return None
-
-    @staticmethod
-    def _scope_details(resource):
-        """Return scope details from either a model resource or raw API dict."""
-        if hasattr(resource, "scope_details"):
-            return getattr(resource, "scope_details", None)
-        if isinstance(resource, dict):
-            return resource.get("scopeDetails")
-        return None
-
-    @staticmethod
-    def _dict_scope_key(attr_name: str) -> str:
-        """Convert a snake_case scope attribute name to the ND API camelCase key."""
-        parts = attr_name.split("_")
-        return parts[0] + "".join(part.title() for part in parts[1:])
-
-    @staticmethod
-    def _extract_scope_switch_key_val(scope_details, switch_key, src_switch_key, log: logging.Logger) -> str | None:
-        """Extract a switch identifier from a scope_details model using the correct attribute name.
-
-        Selects between ``switch_key`` (for single-switch scopes: device, device_interface)
-        and ``src_switch_key`` (for dual-switch scopes: device_pair, link).  Returns None
-        for fabric-scoped resources which carry no switch identity.
-
-        Args:
-            scope_details: A scope model instance (FabricScope, DeviceScope,
-                DeviceInterfaceScope, DevicePairScope, LinkScope) or None.
-            switch_key: Attribute name to read for single-switch scopes
-                (e.g. ``'switch_id'`` or ``'switch_ip'``).
-            src_switch_key: Attribute name to read for dual-switch scopes
-                (e.g. ``'src_switch_id'`` or ``'src_switch_ip'``).
-            log: Logger instance.
-
-        Returns:
-            The switch identifier string, or None if the scope is fabric-level
-            or ``scope_details`` is None.
-        """
-        if scope_details is None:
-            log.debug("_extract_scope_switch_key_val: scope_details is None, returning None")
-            return None
-        if isinstance(scope_details, dict):
-            scope_type = scope_details.get("scopeType")
-            if scope_type == "fabric":
-                log.debug("_extract_scope_switch_key_val: fabric scope dict has no switch identity, returning None")
-                return None
-            switch_dict_key = ResourceManagerDiffEngine._dict_scope_key(switch_key)
-            src_switch_dict_key = ResourceManagerDiffEngine._dict_scope_key(src_switch_key)
-            if scope_type in ("device", "deviceInterface"):
-                value = scope_details.get(switch_dict_key)
-                log.debug("_extract_scope_switch_key_val: dict %s scope, %s='%s'", scope_type, switch_dict_key, value)
-                return value
-            if scope_type in ("devicePair", "link"):
-                value = scope_details.get(src_switch_dict_key)
-                log.debug("_extract_scope_switch_key_val: dict %s scope, %s='%s'", scope_type, src_switch_dict_key, value)
-                return value
-            value = scope_details.get(switch_dict_key) or scope_details.get(src_switch_dict_key)
-            log.debug("_extract_scope_switch_key_val: unknown dict scope type %s, fallback value='%s'", scope_type, value)
-            return value
-        if isinstance(scope_details, FabricScope):
-            log.debug("_extract_scope_switch_key_val: FabricScope has no switch identity, returning None")
-            return None
-        if isinstance(scope_details, (DeviceScope, DeviceInterfaceScope)):
-            value = getattr(scope_details, switch_key, None)
-            log.debug("_extract_scope_switch_key_val: %s scope, %s='%s'", type(scope_details).__name__, switch_key, value)
-            return value
-        if isinstance(scope_details, (DevicePairScope, LinkScope)):
-            value = getattr(scope_details, src_switch_key, None)
-            log.debug("_extract_scope_switch_key_val: %s scope, %s='%s'", type(scope_details).__name__, src_switch_key, value)
-            return value
-        # Fallback: try common attribute names
-        value = getattr(scope_details, switch_key, None) or getattr(scope_details, src_switch_key, None)
-        log.debug("_extract_scope_switch_key_val: unknown scope type %s, fallback value='%s'", type(scope_details).__name__, value)
-        return value
-
-    @staticmethod
-    def _extract_scope_type(scope_details, log: logging.Logger) -> str | None:
-        """Extract and map the playbook-style scope_type from a scope_details model.
-
-        Args:
-            scope_details: A scope model instance.
-            log: Logger instance.
-
-        Returns:
-            Playbook-style scope_type string (e.g. 'device_interface'), or None.
-        """
-        if scope_details is None:
-            log.debug("_extract_scope_type: scope_details is None, returning None")
-            return None
-        if isinstance(scope_details, dict):
-            raw = scope_details.get("scopeType")
-        else:
-            raw = getattr(scope_details, "scope_type", None)
-        if not raw:
-            log.debug("_extract_scope_type: no scope_type attribute on %s, returning None", type(scope_details).__name__)
-            return None
-        mapped = API_SCOPE_TYPE_TO_PLAYBOOK.get(raw, raw)
-        log.debug("_extract_scope_type: raw='%s' mapped to '%s'", raw, mapped)
-        return mapped
-
-    @staticmethod
-    def _compare_resource_values(have: str, want: str, log: logging.Logger) -> bool:
-        """Compare resource values with IPv4/IPv6 network awareness.
-
-        Args:
-            have: Existing resource value from the API.
-            want: Proposed resource value from the playbook.
-            log: Logger instance
-
-        Returns:
-            True if the values are functionally equivalent, False otherwise.
-        """
-        if have is None and want is None:
-            log.debug("_compare_resource_values: both have and want are None, returning True")
-            return True
-        if have is None or want is None:
-            log.debug("_compare_resource_values: one of have or want is None (have=%s, want=%s), returning False", have, want)
-            return False
-
-        have = str(have).strip()
-        want = str(want).strip()
-
-        def _classify(val):
-            if "/" in val:
-                try:
-                    parsed = ipaddress.ip_network(val, strict=False)
-                    log.debug("_compare_resource_values: classified '%s' as network: %s", val, parsed)
-                    return "network", parsed
-                except ValueError:
-                    log.debug("_compare_resource_values: failed to parse '%s' as network, continuing", val)
-            try:
-                parsed = ipaddress.ip_address(val)
-                log.debug("_compare_resource_values: classified '%s' as address: %s", val, parsed)
-                return "address", parsed
-            except ValueError:
-                log.debug("_compare_resource_values: failed to parse '%s' as address, classifying as raw", val)
-                log.debug("_compare_resource_values: classified '%s' as raw string", val)
-            return "raw", val
-
-        th, vh = _classify(have)
-        tw, vw = _classify(want)
-
-        if th == tw == "address":
-            result = vh.exploded == vw.exploded
-            log.debug("_compare_resource_values: both are addresses (have=%s, want=%s), exploded comparison result=%s", vh.exploded, vw.exploded, result)
-            return result
-        if th == tw == "network":
-            result = vh == vw
-            log.debug("_compare_resource_values: both are networks (have=%s, want=%s), comparison result=%s", vh, vw, result)
-            return result
-        result = have == want
-        log.debug("_compare_resource_values: raw string comparison (have='%s', want='%s'), result=%s", have, want, result)
-        return result
-
-    @staticmethod
-    def _make_resource_key(
-        entity_name: str | None,
-        pool_name: str | None,
-        scope_type: str | None,
-        switch_ip: str | None,
-        log: logging.Logger,
-    ) -> tuple:
-        """Build a normalized deduplication key for a resource entry.
-
-        Args:
-            entity_name: Resource entity name (will be tilde-normalized).
-            pool_name: Pool name.
-            scope_type: Playbook-style scope type.
-            switch_ip: Switch IP, or None for fabric-scoped resources.
-            log: Logger instance.
-
-        Returns:
-            Tuple used as a dict key for matching proposed vs existing.
-        """
-        norm_entity = ResourceManagerDiffEngine._normalize_entity_key(entity_name, log=log) if entity_name else None
-        log.debug("_make_resource_key: entity_name provided %s, normalized to '%s'", entity_name, norm_entity)
-
-        norm_pool = ResourceManagerDiffEngine._normalize_pool_name(pool_name, log=log)
-        log.debug("_make_resource_key: pool_name='%s' normalized to '%s'", pool_name, norm_pool)
-
-        # device_pair and link encode both endpoints in entity_name;
-        # normalize switch to None so existing_index and proposed lookups align.
-        if scope_type in ("device_pair", "link"):
-            norm_switch = None
-            log.debug("_make_resource_key: scope_type='%s' is multi-endpoint, setting norm_switch=None (original switch_ip='%s')", scope_type, switch_ip)
-        else:
-            norm_switch = switch_ip
-            log.debug("_make_resource_key: scope_type='%s' is single-endpoint, keeping norm_switch='%s'", scope_type, norm_switch)
-
-        key = (norm_entity, norm_pool, scope_type, norm_switch)
-        log.debug(
-            "_make_resource_key: built key=%s from entity_name='%s', pool_name='%s', scope_type='%s', switch_ip='%s'",
-            key,
-            entity_name,
-            pool_name,
-            scope_type,
-            switch_ip,
-        )
-
-        return key
-
-    @staticmethod
-    def validate_configs(
-        config: dict[str, Any] | list[dict[str, Any]],
-        state: str,
-        log: logging.Logger,
-    ) -> list[ResourceManagerConfigModel]:
-        """Validate raw module config and return typed resource configurations.
-
-        Args:
-            config: Raw config dict or list of dicts from module parameters.
-            state: Requested module state.
-            log: Logger instance.
-
-        Returns:
-            list of validated ``ResourceManagerConfigModel`` objects.
-        """
-        log.debug("ENTER: validate_configs()")
-
-        configs_list = config if isinstance(config, list) else [config]
-        log.debug("Normalized to %s configuration(s)", len(configs_list))
-
-        validated_configs: list[ResourceManagerConfigModel] = []
-        for idx, cfg in enumerate(configs_list):
-            try:
-                validated = ResourceManagerConfigModel.model_validate(cfg, context={"state": state})
-                validated_configs.append(validated)
-            except ValidationError as e:
-                error_detail = e.errors() if hasattr(e, "errors") else str(e)
-                error_msg = f"Configuration validation failed for config index {idx}: {error_detail}"
-                log.error(error_msg)
-                raise ValueError(error_msg) from e
-            except Exception as e:
-                error_msg = f"Configuration validation failed for config index {idx}: {str(e)}"
-                log.error(error_msg)
-                raise ValueError(error_msg) from e
-
-        if not validated_configs:
-            log.warning("No valid configurations found in input")
-            return validated_configs
-
-        # Duplicate check: (entity_name, pool_name, scope_type, frozenset(switch))
-        seen_keys: set = set()
-        duplicate_keys: set = set()
-        log.debug(
-            "validate_configs: starting duplicate check on %s validated config(s)",
-            len(validated_configs),
-        )
-        for cfg_dup_idx, cfg in enumerate(validated_configs):
-            key = (
-                cfg.entity_name,
-                cfg.pool_name,
-                cfg.scope_type,
-                frozenset(cfg.switches or []),
-            )
-            log.debug(
-                "validate_configs: duplicate-check [%s] — entity_name='%s', pool_name='%s', scope_type='%s', switches=%s, key_seen_before=%s",
-                cfg_dup_idx,
-                cfg.entity_name,
-                cfg.pool_name,
-                cfg.scope_type,
-                list(cfg.switches or []),
-                key in seen_keys,
-            )
-            if key in seen_keys:
-                log.warning(
-                    "validate_configs: [%s] duplicate key detected — entity_name='%s', pool_name='%s', scope_type='%s'",
-                    cfg_dup_idx,
-                    cfg.entity_name,
-                    cfg.pool_name,
-                    cfg.scope_type,
-                )
-                duplicate_keys.add(key)
-            else:
-                log.debug(
-                    "validate_configs: [%s] key is unique so far — entity_name='%s'",
-                    cfg_dup_idx,
-                    cfg.entity_name,
-                )
-            seen_keys.add(key)
-
-        if duplicate_keys:
-            error_msg = f"Duplicate config entries found: {[str(k) for k in duplicate_keys]}. Each resource must appear only once."
-            log.error(error_msg)
-            raise ValueError(error_msg)
-
-        log.info(
-            "Successfully validated %s configuration(s)",
-            len(validated_configs),
-        )
-        log.debug("EXIT: validate_configs() -> %s configs", len(validated_configs))
-        return validated_configs
-
-    @staticmethod
-    def compute_changes(
-        proposed: list[ResourceManagerConfigModel],
-        existing: list[ResourceManagerResponse],
-        log: logging.Logger,
-    ) -> dict[str, list]:
-        """Compare proposed and existing resources and categorize changes.
-
-        Uses ``ResourceManagerResponse`` fields (``entity_name``, ``pool_name``,
-        ``scope_details``, ``resource_value``) to build a matching index and
-        classify each proposed entry.
-
-        Args:
-            proposed: Validated ``ResourceManagerConfigModel`` objects
-                representing desired state.
-            existing: ``ResourceManagerResponse`` models from the ND API
-                representing current state.
-            log: Logger instance.
-
-        Returns:
-            dict mapping change buckets to item lists:
-              - ``to_add``:     ``(ResourceManagerConfigModel, switch_ip)`` tuples
-              - ``to_update``:  ``(ResourceManagerConfigModel, switch_ip)`` tuples
-              - ``to_delete``:  ``ResourceManagerResponse`` items
-              - ``idempotent``: ``(ResourceManagerConfigModel, switch_ip)`` tuples
-        """
-        log.debug("ENTER: compute_changes()")
-        log.debug(
-            "Comparing %s proposed vs %s existing resources",
-            len(proposed),
-            len(existing),
-        )
-
-        # Build index of existing resources keyed by
-        # (normalized_entity, pool_name, playbook_scope_type, switch_id)
-        existing_index: dict[tuple, ResourceManagerResponse] = {}
-        for res in existing:
-            entity = ResourceManagerDiffEngine._resource_attr(res, "entity_name", "entityName")
-            pool = ResourceManagerDiffEngine._resource_attr(res, "pool_name", "poolName")
-            scope_details = ResourceManagerDiffEngine._scope_details(res)
-            scope_type = ResourceManagerDiffEngine._extract_scope_type(scope_details, log=log)
-            switch_id = ResourceManagerDiffEngine._extract_scope_switch_key_val(
-                scope_details, switch_key="switch_id", src_switch_key="src_switch_id", log=log
-            )
-            key = ResourceManagerDiffEngine._make_resource_key(entity, pool, scope_type, switch_id, log=log)
-            existing_index[key] = res
-            log.debug(
-                "Existing index entry: entity=%s, pool=%s, scope_type=%s, switch_id=%s",
-                entity,
-                pool,
-                scope_type,
-                switch_id,
-            )
-
-        log.debug("Built existing index with %s entries", len(existing_index))
-
-        changes: dict[str, list] = {
-            "to_add": [],
-            "to_update": [],
-            "to_delete": [],
-            "idempotent": [],
-            "debugs": [],
-        }
-
-        # Build a secondary index keyed by normalised entity_name only.
-        # Used to detect partial matches (same entity, different pool/scope/switch)
-        # and populate the debugs bucket to mirror ND's mismatch logging.
-        entity_only_index: dict[str, list[ResourceManagerResponse]] = {}
-        for res in existing:
-            entity_name = ResourceManagerDiffEngine._resource_attr(res, "entity_name", "entityName") or ""
-            norm = ResourceManagerDiffEngine._normalize_entity_key(entity_name, log=log)
-            entity_only_index.setdefault(norm, []).append(res)
-            log.debug(
-                "entity_only_index: added entity='%s' under norm_key='%s' (total under key: %s)",
-                entity_name,
-                norm,
-                len(entity_only_index[norm]),
-            )
-
-        log.debug("Built entity_only_index with %s unique normalised key(s)", len(entity_only_index))
-
-        # Track which existing keys matched at least one proposed entry
-        matched_existing_keys: set = set()
-        # Track partial-match diagnostics already emitted to avoid duplicates.
-        seen_mismatch_keys: set = set()
-
-        # Categorise proposed resources
-        for cfg in proposed:
-            scope_type = cfg.scope_type
-            pool_name = cfg.pool_name
-            entity_name = cfg.entity_name
-            resource_value = cfg.resource
-
-            log.debug(
-                "Processing proposed cfg: entity=%s, pool=%s, scope=%s, resource=%s, switches=%s",
-                entity_name,
-                pool_name,
-                scope_type,
-                resource_value,
-                cfg.switches,
-            )
-
-            # device_pair and link encode both endpoints in entity_name; one lookup covers the pair.
-            if scope_type in ("device_pair", "link"):
-                switches = [None]
-                log.debug(
-                    "scope_type='%s' is multi-endpoint — using single switch=None lookup for entity='%s'",
-                    scope_type,
-                    entity_name,
-                )
-            else:
-                switches = cfg.switches if (scope_type != "fabric" and cfg.switches) else [None]
-                log.debug(
-                    "scope_type='%s' — resolved switches=%s for entity='%s'",
-                    scope_type,
-                    switches,
-                    entity_name,
-                )
-
-            for sw in switches:
-                key = ResourceManagerDiffEngine._make_resource_key(entity_name, pool_name, scope_type, sw, log=log)
-                log.debug(
-                    "Lookup key=%s for entity='%s', pool='%s', scope='%s', switch=%s",
-                    key,
-                    entity_name,
-                    pool_name,
-                    scope_type,
-                    sw,
-                )
-                existing_res = existing_index.get(key)
-
-                if existing_res is None:
-                    log.info(
-                        "Resource (entity=%s, pool=%s, scope=%s, switch=%s) not found in existing — marking to_add",
-                        entity_name,
-                        pool_name,
-                        scope_type,
-                        sw,
-                    )
-                    changes["to_add"].append((cfg, sw, None))
-
-                    # GAP-7: Partial-match detection — same entity_name, different
-                    # pool_name / scope_type / switch_ip.  Mirrors ND's
-                    # nd_rm_get_mismatched_values() / changed_dict["debugs"] logic.
-                    norm = ResourceManagerDiffEngine._normalize_entity_key(entity_name, log=log)
-                    partials = entity_only_index.get(norm, [])
-                    log.debug(
-                        "Partial-match scan for entity='%s' (norm='%s'): %s candidate(s)",
-                        entity_name,
-                        norm,
-                        len(partials),
-                    )
-                    for partial in partials:
-                        partial_resource_id = ResourceManagerDiffEngine._resource_attr(partial, "resource_id", "resourceId")
-                        mismatch_key = (
-                            entity_name,
-                            partial_resource_id,
-                        )
-                        if mismatch_key in seen_mismatch_keys:
-                            log.debug(
-                                "compute_changes: skipping duplicate partial match for entity='%s', resource_id=%s",
-                                entity_name,
-                                ResourceManagerDiffEngine._resource_attr(partial, "resource_id", "resourceId"),
-                            )
-                            continue
-                        seen_mismatch_keys.add(mismatch_key)
-
-                        partial_scope_details = ResourceManagerDiffEngine._scope_details(partial)
-                        partial_pool = ResourceManagerDiffEngine._normalize_pool_name(
-                            ResourceManagerDiffEngine._resource_attr(partial, "pool_name", "poolName"),
-                            log=log,
-                        )
-                        desired_pool = ResourceManagerDiffEngine._normalize_pool_name(pool_name, log=log)
-                        partial_scope = ResourceManagerDiffEngine._extract_scope_type(partial_scope_details, log=log)
-                        partial_sw = ResourceManagerDiffEngine._extract_scope_switch_key_val(
-                            partial_scope_details, switch_key="switch_ip", src_switch_key="src_switch_ip", log=log
-                        )
-                        partial_resource_value = ResourceManagerDiffEngine._resource_attr(partial, "resource_value", "resourceValue")
-                        existing_values = {
-                            "resource_id": partial_resource_id,
-                            "pool_name": partial_pool,
-                            "scope_type": partial_scope,
-                            "switch_ip": partial_sw,
-                            "resource_value": partial_resource_value,
-                        }
-                        mismatch = {
-                            "resource_id": partial_resource_id,
-                            "have_pool_name": partial_pool,
-                            "want_pool_name": desired_pool,
-                            "have_scope_type": partial_scope,
-                            "want_scope_type": scope_type,
-                            "have_switch_ip": partial_sw,
-                            "have_resource_value": partial_resource_value,
-                            "want_resource_value": resource_value,
-                        }
-                        log.debug(
-                            "compute_changes: partial match for entity='%s': existing=%s mismatch=%s",
-                            entity_name,
-                            existing_values,
-                            mismatch,
-                        )
-                        changes["debugs"].append(
-                            {
-                                "Entity Name": entity_name,
-                                "EXISTING_VALUES": existing_values,
-                                "MISMATCHED_VALUES": mismatch,
-                            }
-                        )
-                else:
-                    log.debug(
-                        "Resource (entity=%s, pool=%s, scope=%s, switch=%s) found in existing — resource_id=%s, existing_value='%s'",
-                        entity_name,
-                        pool_name,
-                        scope_type,
-                        sw,
-                        ResourceManagerDiffEngine._resource_attr(existing_res, "resource_id", "resourceId"),
-                        ResourceManagerDiffEngine._resource_attr(existing_res, "resource_value", "resourceValue"),
-                    )
-                    matched_existing_keys.add(key)
-                    existing_value = ResourceManagerDiffEngine._resource_attr(existing_res, "resource_value", "resourceValue")
-
-                    if ResourceManagerDiffEngine._compare_resource_values(existing_value, resource_value, log=log):
-                        log.debug(
-                            "Resource (entity=%s, pool=%s, scope=%s, switch=%s) is idempotent (value=%s)",
-                            entity_name,
-                            pool_name,
-                            scope_type,
-                            sw,
-                            existing_value,
-                        )
-                        changes["idempotent"].append((cfg, sw, existing_res))
-                    else:
-                        log.info(
-                            "Resource (entity=%s, pool=%s, scope=%s, switch=%s) value differs (existing=%s, desired=%s) — marking to_update",
-                            entity_name,
-                            pool_name,
-                            scope_type,
-                            sw,
-                            existing_value,
-                            resource_value,
-                        )
-                        changes["to_update"].append((cfg, sw, existing_res))
-
-        log.debug(
-            "Proposed scan complete — matched_existing_keys=%s, total existing_index keys=%s",
-            len(matched_existing_keys),
-            len(existing_index),
-        )
-
-        # Resources in existing but not matched by any proposed entry → to_delete
-        for key, res in existing_index.items():
-            if key not in matched_existing_keys:
-                log.info(
-                    "Existing resource (entity=%s, pool=%s) not in proposed — marking to_delete",
-                    ResourceManagerDiffEngine._resource_attr(res, "entity_name", "entityName"),
-                    ResourceManagerDiffEngine._resource_attr(res, "pool_name", "poolName"),
-                )
-                changes["to_delete"].append(res)
-            else:
-                log.debug(
-                    "Existing resource (entity=%s, pool=%s, key=%s) was matched by a proposed entry — skipping to_delete",
-                    ResourceManagerDiffEngine._resource_attr(res, "entity_name", "entityName"),
-                    ResourceManagerDiffEngine._resource_attr(res, "pool_name", "poolName"),
-                    key,
-                )
-
-        log.info(
-            "Compute changes summary: to_add=%s, to_update=%s, to_delete=%s, idempotent=%s, debugs=%s",
-            len(changes["to_add"]),
-            len(changes["to_update"]),
-            len(changes["to_delete"]),
-            len(changes["idempotent"]),
-            len(changes["debugs"]),
-        )
-        log.debug("EXIT: compute_changes()")
-        return changes
-
-    @staticmethod
-    def validate_resource_api_fields(
-        nd: NDModule,
-        resource_cfg: ResourceManagerConfigModel,
-        api_resource: ResourceManagerResponse,
-        context: str,
-        log: logging.Logger,
-    ) -> None:
-        """Validate user-supplied resource fields against the ND API response.
-
-        Only fields that are non-None in ``resource_cfg`` are validated.
-        Fields omitted by the user are silently accepted from the API response.
-        Uses ``ResourceManagerResponse`` model attributes directly for
-        field access (``entity_name``, ``pool_name``, ``resource_value``,
-        ``scope_details``).
-
-        Args:
-            nd: ND module wrapper used for failure handling.
-            resource_cfg: Validated resource config from the playbook.
-            api_resource: Matching ``ResourceManagerResponse`` from the ND API.
-            log: Logger instance.
-            context: Label used in error messages (e.g. ``"Resource"``).
-
-        Returns:
-            None.
-
-        Raises:
-            ValueError: When any provided field does not match the API response.
-        """
-        mismatches: list[str] = []
-
-        # entity_name: tilde-order-insensitive comparison
-        if resource_cfg.entity_name is not None:
-            cfg_norm = ResourceManagerDiffEngine._normalize_entity_key(resource_cfg.entity_name, log=log)
-            api_norm = ResourceManagerDiffEngine._normalize_entity_key(api_resource.entity_name, log=log) if api_resource.entity_name else None
-
-            log.debug(
-                "validate_resource_api_fields: checking entity_name — cfg_norm='%s', api_norm='%s'",
-                cfg_norm,
-                api_norm,
-            )
-            if cfg_norm != api_norm:
-                log.debug(
-                    "validate_resource_api_fields: entity_name MISMATCH — provided='%s', API='%s'",
-                    resource_cfg.entity_name,
-                    api_resource.entity_name,
-                )
-                mismatches.append(f"entity_name: provided '{resource_cfg.entity_name}', API reports '{api_resource.entity_name}'")
-            else:
-                log.debug(
-                    "validate_resource_api_fields: entity_name OK — '%s' matches API",
-                    resource_cfg.entity_name,
-                )
-        else:
-            log.debug(
-                "validate_resource_api_fields: entity_name not provided in cfg — skipping check (api_entity_name='%s')",
-                api_resource.entity_name,
-            )
-
-        # pool_name: exact match
-        if resource_cfg.pool_name is not None:
-            cfg_pool_norm = ResourceManagerDiffEngine._normalize_pool_name(resource_cfg.pool_name, log=log)
-            api_pool_norm = ResourceManagerDiffEngine._normalize_pool_name(api_resource.pool_name, log=log)
-            log.debug(
-                "validate_resource_api_fields: checking pool_name — cfg='%s' (norm='%s'), api='%s' (norm='%s')",
-                resource_cfg.pool_name,
-                cfg_pool_norm,
-                api_resource.pool_name,
-                api_pool_norm,
-            )
-            if cfg_pool_norm != api_pool_norm:
-                log.debug(
-                    "validate_resource_api_fields: pool_name MISMATCH — provided='%s', API='%s'",
-                    resource_cfg.pool_name,
-                    api_resource.pool_name,
-                )
-                mismatches.append(f"pool_name: provided '{resource_cfg.pool_name}', API reports '{api_resource.pool_name}'")
-            else:
-                log.debug(
-                    "validate_resource_api_fields: pool_name OK — '%s' matches API",
-                    resource_cfg.pool_name,
-                )
-        else:
-            log.debug(
-                "validate_resource_api_fields: pool_name not provided in cfg — skipping check (api_pool_name='%s')",
-                api_resource.pool_name,
-            )
-
-        # resource vs resource_value: IPv4/v6-aware comparison
-        if resource_cfg.resource is not None:
-            log.debug(
-                "validate_resource_api_fields: checking resource value — cfg='%s', api='%s'",
-                resource_cfg.resource,
-                api_resource.resource_value,
-            )
-            if not ResourceManagerDiffEngine._compare_resource_values(api_resource.resource_value, resource_cfg.resource, log=log):
-                log.debug(
-                    "validate_resource_api_fields: resource value MISMATCH — provided='%s', API='%s'",
-                    resource_cfg.resource,
-                    api_resource.resource_value,
-                )
-                mismatches.append(f"resource: provided '{resource_cfg.resource}', API reports '{api_resource.resource_value}'")
-            else:
-                log.debug(
-                    "validate_resource_api_fields: resource value OK — '%s' matches API '%s'",
-                    resource_cfg.resource,
-                    api_resource.resource_value,
-                )
-        else:
-            log.debug(
-                "validate_resource_api_fields: resource not provided in cfg — skipping check (api_resource_value='%s')",
-                api_resource.resource_value,
-            )
-
-        if mismatches:
-            raise ValueError(
-                f"{context} field mismatch for entity '{resource_cfg.entity_name}'. "
-                f"The following provided values do not match the API data:\n\n".join(f"  - {m}" for m in mismatches)
-            )
-
-        log.debug(
-            "validate_resource_api_fields: all provided fields match API for entity='%s', pool='%s'",
-            resource_cfg.entity_name,
-            resource_cfg.pool_name,
-        )
-
-
-class NDResourceManagerModule:
+class NDResourceManagerModule(ResourceManagerResourceHelpersMixin):
     """
     Manage resources in Cisco Nexus Dashboard via the ND Manage v1 API.
 
@@ -1207,434 +424,6 @@ class NDResourceManagerModule:
         )
         return config_copy
 
-    # ------------------------------------------------------------------
-    # Resource attribute accessors (handle both ResourceManagerResponse and raw dict)
-    # ------------------------------------------------------------------
-
-    def _attr(self, resource, model_attr, dict_key):
-        """Return a field value from a resource that may be a model instance or a raw dict.
-
-        Delegates to ``ResourceManagerDiffEngine._resource_attr`` so diffing,
-        gathered output, and filtering share the same model/dict access behavior.
-
-        Args:
-            resource: A ``ResourceManagerResponse`` model instance or a plain dict.
-            model_attr: Attribute name to access on a model instance (snake_case).
-            dict_key: Key to access on a raw dict (camelCase, e.g. ``'entityName'``).
-
-        Returns:
-            The field value, or None if neither path resolves.
-        """
-        value = ResourceManagerDiffEngine._resource_attr(resource, model_attr, dict_key)
-        self.log.debug("_attr: resolved '%s'/'%s' from %s: %s", model_attr, dict_key, type(resource).__name__, value)
-        return value
-
-    def _get_entity_name(self, resource):
-        """Return the entity_name field from a resource model or raw dict."""
-        return self._attr(resource, "entity_name", "entityName")
-
-    def _get_pool_name(self, resource):
-        """Return the pool_name field from a resource model or raw dict."""
-        return self._attr(resource, "pool_name", "poolName")
-
-    def _get_resource_id(self, resource):
-        """Return the resource_id field from a resource model or raw dict."""
-        return self._attr(resource, "resource_id", "resourceId")
-
-    def _get_resource_value(self, resource):
-        """Return the resource_value field from a resource model or raw dict."""
-        return self._attr(resource, "resource_value", "resourceValue")
-
-    def _get_scope_type(self, resource):
-        """Return the playbook-style scope_type string for a resource.
-
-        Reads the raw ND API ``scopeType`` value from either the model's
-        ``scope_details.scope_type`` attribute or the ``scopeDetails.scopeType`` key of a
-        raw dict, then maps it from the API camelCase format (e.g. ``'deviceInterface'``)
-        to the playbook format (e.g. ``'device_interface'``) using
-        ``API_SCOPE_TYPE_TO_PLAYBOOK``.
-
-        Args:
-            resource: A ``ResourceManagerResponse`` model instance or a plain dict.
-
-        Returns:
-            Playbook-style scope_type string, or None if the resource type is unrecognised.
-        """
-        scope_details = ResourceManagerDiffEngine._scope_details(resource)
-        if scope_details is None:
-            self.log.debug("_get_scope_type: unrecognised resource type %s, returning None", type(resource))
-            return None
-        mapped = ResourceManagerDiffEngine._extract_scope_type(scope_details, log=self.log)
-        self.log.debug("_get_scope_type: mapped scope to playbook scope '%s'", mapped)
-        return mapped
-
-    def _get_switch_ip(self, resource):
-        """Return the primary switch IP/ID from scopeDetails (src switch for device_pair/link).
-
-        Delegates to ResourceManagerDiffEngine._extract_scope_switch_key_val for model
-        instances so that all scope types are handled uniformly:
-          - fabric              → None
-          - device / device_interface → switch_ip
-          - device_pair / link  → src_switch_ip
-        """
-        scope_details = ResourceManagerDiffEngine._scope_details(resource)
-        if scope_details is not None:
-            value = ResourceManagerDiffEngine._extract_scope_switch_key_val(
-                scope_details, switch_key="switch_ip", src_switch_key="src_switch_ip", log=self.log
-            )
-            self.log.debug("_get_switch_ip: from scope_details, switch_ip=%s", value)
-            return value
-        self.log.debug("_get_switch_ip: unrecognised resource type %s, returning None", type(resource))
-        return None
-
-    # ------------------------------------------------------------------
-    # Matching helpers
-    # ------------------------------------------------------------------
-
-    def _entity_names_match(self, e1, e2):
-        """Compare two entity names in a tilde-order-insensitive way.
-
-        Splits each name on ``'~'``, sorts the resulting parts alphabetically, and
-        compares the sorted lists.  This ensures that a device_pair entity such as
-        ``'SER1~SER2~label'`` matches ``'SER2~SER1~label'`` regardless of the order
-        in which the serial numbers appear in the playbook vs the ND API response.
-
-        Args:
-            e1: First entity name string.
-            e2: Second entity name string.
-
-        Returns:
-            True if both names are non-None and their sorted tilde-parts are equal,
-            False otherwise.
-        """
-        if e1 is None or e2 is None:
-            self.log.debug(
-                "_entity_names_match: one or both entity names are None (e1=%s, e2=%s), returning False",
-                e1,
-                e2,
-            )
-            return False
-        result = sorted(e1.split("~")) == sorted(e2.split("~"))
-        self.log.debug(
-            "_entity_names_match: e1='%s', e2='%s', sorted_e1=%s, sorted_e2=%s, match=%s",
-            e1,
-            e2,
-            sorted(e1.split("~")),
-            sorted(e2.split("~")),
-            result,
-        )
-        return result
-
-    # ------------------------------------------------------------------
-    # API payload builders
-    # ------------------------------------------------------------------
-
-    def _build_fabric_scope(self, _switch_id=None, _entity_name=None):
-        """Build fabric-level scope details."""
-        return FabricScope(fabric_name=self.fabric)
-
-    def _build_device_scope(self, switch_id=None, _entity_name=None):
-        """Build device-level scope details."""
-        return DeviceScope(switch_id=switch_id)
-
-    def _build_device_interface_scope(self, switch_id=None, entity_name=None):
-        """Build device-interface scope details from ``entity_name``."""
-        parts = (entity_name or "").split("~", 1)
-        interface_name = parts[1] if len(parts) > 1 else None
-        if not interface_name:
-            self.log.warning(
-                "_build_device_interface_scope: could not parse interfaceName from entity_name='%s'",
-                entity_name,
-            )
-        return DeviceInterfaceScope(switch_id=switch_id, interface_name=interface_name)
-
-    def _build_device_pair_scope(self, _switch_id=None, entity_name=None):
-        """Build device-pair scope details from ``entity_name``."""
-        parts = (entity_name or "").split("~")
-        src_switch_id = parts[0] if len(parts) > 0 else None
-        dst_switch_id = parts[1] if len(parts) > 1 else None
-        return DevicePairScope(src_switch_id=src_switch_id, dst_switch_id=dst_switch_id)
-
-    def _build_link_scope(self, _switch_id=None, entity_name=None):
-        """Build link scope details from ``entity_name``."""
-        parts = (entity_name or "").split("~")
-        src_switch_id = parts[0] if len(parts) > 0 else None
-        src_interface_name = parts[1] if len(parts) > 1 else None
-        dst_switch_id = parts[2] if len(parts) > 2 else None
-        dst_interface_name = parts[3] if len(parts) > 3 else None
-        return LinkScope(
-            src_switch_id=src_switch_id,
-            src_interface_name=src_interface_name,
-            dst_switch_id=dst_switch_id,
-            dst_interface_name=dst_interface_name,
-        )
-
-    def _build_scope_details(self, scope_type, switch_ip=None, entity_name=None):
-        """Build the scopeDetails model for the ND Manage API.
-
-        ``switch_ip`` is the translated switchId (serial number) of the source switch
-        from the playbook ``switch`` list.  The entity_name encodes the full topology
-        (src and dst) as tilde-separated fields for multi-switch scopes.
-
-          - fabric:           FabricScope(fabricName)
-          - device:           DeviceScope(switchId)
-          - device_interface: DeviceInterfaceScope(switchId, interfaceName)
-          - device_pair:      DevicePairScope(srcSwitchId, dstSwitchId)
-          - link:             LinkScope(src/dst switch and interface details)
-        """
-        self.log.debug(
-            "_build_scope_details: scope_type=%s, switch_ip=%s, entity_name=%s, fabric=%s",
-            scope_type,
-            switch_ip,
-            entity_name,
-            self.fabric,
-        )
-
-        scope_builders = {
-            "fabric": self._build_fabric_scope,
-            "device": self._build_device_scope,
-            "device_interface": self._build_device_interface_scope,
-            "device_pair": self._build_device_pair_scope,
-            "link": self._build_link_scope,
-        }
-        builder = scope_builders.get(scope_type)
-        if builder is None:
-            raise ValueError("Unsupported scope_type '{0}' while building scope details".format(scope_type))
-
-        result = builder(switch_ip, entity_name)
-        self.log.debug("_build_scope_details: result=%s", result)
-        return result
-
-    def _build_create_payload(self, cfg, switch_ip=None):
-        """Build the POST body for a single resource creation request.
-
-        Accepts either a typed ``ResourceManagerConfigModel`` instance or a legacy dict
-        (backward-compatible path).  Delegates scope construction to
-        ``_build_scope_details`` and serialises the complete request via
-        ``ResourceManagerRequest.to_payload()``.
-
-        Args:
-            cfg: A ``ResourceManagerConfigModel`` instance or a dict with keys
-                ``scope_type``, ``entity_name``, ``pool_name``, ``pool_type``,
-                and optionally ``resource``.
-            switch_ip: The resolved switchId (serial number) for the primary switch,
-                or None for fabric-scoped resources.
-
-        Returns:
-            A plain dict payload ready to be sent to the ND Manage API POST endpoint.
-        """
-        if isinstance(cfg, ResourceManagerConfigModel):
-            scope_type = cfg.scope_type
-            entity_name = cfg.entity_name
-            pool_name = cfg.pool_name
-            pool_type = cfg.pool_type
-            resource_value = cfg.resource
-        else:
-            # Legacy dict path (kept for backward-compat with any callers not yet refactored)
-            scope_type = cfg["scope_type"]
-            entity_name = cfg["entity_name"]
-            pool_name = cfg["pool_name"]
-            pool_type = cfg.get("pool_type")
-            resource_value = cfg.get("resource")
-
-        self.log.debug(
-            "_build_create_payload: pool_name=%s, pool_type=%s, entity_name=%s, scope_type=%s, switch_ip=%s, resource=%s",
-            pool_name,
-            pool_type,
-            entity_name,
-            scope_type,
-            switch_ip,
-            resource_value,
-        )
-
-        scope = self._build_scope_details(scope_type, switch_ip, entity_name=entity_name)
-
-        request = ResourceManagerRequest(
-            pool_name=pool_name,
-            pool_type=pool_type,
-            entity_name=entity_name,
-            scope_details=scope,
-            is_pre_allocated=True,
-            resource_value=str(resource_value) if resource_value is not None else None,
-        )
-
-        if resource_value is not None:
-            self.log.debug(
-                "_build_create_payload: adding resourceValue='%s' to payload",
-                resource_value,
-            )
-        else:
-            self.log.debug("_build_create_payload: no resource value provided, omitting resourceValue field")
-
-        payload = request.to_payload()
-        self.log.debug("_build_create_payload: final payload=%s", payload)
-        return payload
-
-    # ------------------------------------------------------------------
-    # Gathered results translation
-    # ------------------------------------------------------------------
-
-    def _determine_pool_type(self, resource_value):
-        """Infer the pool_type from a resource value string.
-
-        Attempts to parse the value as an IP network (returns ``'SUBNET'``), then as an
-        IP address (returns ``'IP'``), and falls back to ``'ID'`` for plain integer or
-        string identifiers.  Used when translating raw API responses back into the playbook
-        config format during gathered-state output.
-
-        Args:
-            resource_value: The raw resource value string from the ND API response,
-                e.g. ``'101'``, ``'10.1.1.1'``, or ``'10.1.1.0/24'``.  May be None.
-
-        Returns:
-            One of ``'ID'``, ``'IP'``, or ``'SUBNET'``.
-        """
-        self.log.debug(
-            "_determine_pool_type: evaluating resource_value='%s'",
-            resource_value,
-        )
-        if not resource_value:
-            self.log.debug("_determine_pool_type: resource_value is None/empty — returning 'ID'")
-            return "ID"
-        val = str(resource_value).strip()
-        if "/" in val:
-            self.log.debug(
-                "_determine_pool_type: value='%s' contains '/' — attempting ip_network parse",
-                val,
-            )
-            try:
-                ipaddress.ip_network(val, strict=False)
-                self.log.debug(
-                    "_determine_pool_type: '%s' is a valid IP network — returning 'SUBNET'",
-                    val,
-                )
-                return "SUBNET"
-            except ValueError:
-                self.log.debug(
-                    "_determine_pool_type: '%s' failed ip_network parse — falling through to ip_address check",
-                    val,
-                )
-        else:
-            self.log.debug(
-                "_determine_pool_type: value='%s' has no '/' — skipping ip_network check",
-                val,
-            )
-        try:
-            ipaddress.ip_address(val)
-            self.log.debug(
-                "_determine_pool_type: '%s' is a valid IP address — returning 'IP'",
-                val,
-            )
-            return "IP"
-        except ValueError:
-            self.log.debug(
-                "_determine_pool_type: '%s' is not an IP address — returning 'ID'",
-                val,
-            )
-        return "ID"
-
-    def translate_gathered_results(self, resources):
-        """Translate raw API resource items to the merged-state config format.
-
-        Converts each resource from the ND API response shape
-        (camelCase keys, nested scopeDetails) into the playbook ``config``
-        format used by ``state: merged``:
-          entity_name, pool_type, pool_name, scope_type, resource[, switches].
-        """
-        translated = []
-        self.log.debug(
-            "translate_gathered_results: translating %s resource(s) to playbook config format",
-            len(resources),
-        )
-        for res_idx, res in enumerate(resources):
-            entity_name = self._get_entity_name(res)
-            pool_name = self._get_pool_name(res)
-            resource_value = self._get_resource_value(res)
-            scope_type = self._get_scope_type(res)
-            switch_ip = self._get_switch_ip(res)
-            pool_type = self._determine_pool_type(resource_value)
-            self.log.debug(
-                "translate_gathered_results: [%s] resolved fields — "
-                "entity_name='%s', pool_name='%s', scope_type='%s', "
-                "pool_type='%s', resource_value='%s', switch_ip='%s'",
-                res_idx,
-                entity_name,
-                pool_name,
-                scope_type,
-                pool_type,
-                resource_value,
-                switch_ip,
-            )
-
-            item = {
-                "entity_name": entity_name,
-                "pool_type": pool_type,
-                "pool_name": pool_name,
-                "scope_type": scope_type,
-                "resource": resource_value,
-            }
-            if scope_type != "fabric" and switch_ip:
-                item["switches"] = [switch_ip]
-                self.log.debug(
-                    "translate_gathered_results: [%s] entity='%s' — non-fabric scope ('%s'), adding switches=['%s'] to item",
-                    res_idx,
-                    entity_name,
-                    scope_type,
-                    switch_ip,
-                )
-            else:
-                self.log.debug(
-                    "translate_gathered_results: [%s] entity='%s' — scope_type='%s', switch_ip='%s' — no switch field added",
-                    res_idx,
-                    entity_name,
-                    scope_type,
-                    switch_ip,
-                )
-
-            translated.append(item)
-            self.log.debug(
-                "translate_gathered_results: [%s] appended item=%s",
-                res_idx,
-                item,
-            )
-        self.log.debug(
-            "translate_gathered_results: completed — %s item(s) translated (before switch merge)",
-            len(translated),
-        )
-
-        # Merge entries that share the same (entity_name, pool_name, pool_type,
-        # scope_type, resource) key — only their switch IPs differ.  Fabric-scoped
-        # resources (no 'switches' key) are passed through unchanged.
-        merged: dict = {}
-        for item in translated:
-            key = (
-                item.get("entity_name"),
-                item.get("pool_name"),
-                item.get("pool_type"),
-                item.get("scope_type"),
-                item.get("resource"),
-            )
-            if key in merged:
-                # Accumulate switch IPs for matching entries (deduplicate, preserve order)
-                sw_list = item.get("switches") or []
-                for sw in sw_list:
-                    if sw not in merged[key].get("switches", []):
-                        merged[key]["switches"].append(sw)
-                        self.log.debug(
-                            "translate_gathered_results: merged switch ip='%s' into existing entry for key=%s",
-                            sw,
-                            key,
-                        )
-            else:
-                merged[key] = item
-
-        translated = list(merged.values())
-        self.log.debug(
-            "translate_gathered_results: after switch merge — %s item(s) returned",
-            len(translated),
-        )
-        return translated
-
     def manage_merged(self):
         """Create or update resources to match the desired state defined in the playbook.
 
@@ -1885,112 +674,6 @@ class NDResourceManagerModule:
             path=ep.path,
             payload=remove_req.to_payload(),
         )
-
-    def _get_switch_id(self, resource):
-        """Return the primary switch ID from scopeDetails for gathered filtering."""
-        scope_details = ResourceManagerDiffEngine._scope_details(resource)
-        if scope_details is None:
-            return None
-        return ResourceManagerDiffEngine._extract_scope_switch_key_val(
-            scope_details,
-            switch_key="switch_id",
-            src_switch_key="src_switch_id",
-            log=self.log,
-        )
-
-    @staticmethod
-    def _filter_has_active_criteria(filter_item):
-        """Return True when a gathered filter item has at least one criterion."""
-        return bool(filter_item.get("entity_name") or filter_item.get("pool_name") or filter_item.get("switches"))
-
-    def _resource_matches_filter(self, resource, filter_item):
-        """Return True when a resource matches one gathered filter item."""
-        resource_id = self._get_resource_id(resource)
-        resource_entity = self._get_entity_name(resource)
-        resource_pool = self._get_pool_name(resource)
-        resource_switch_id = self._get_switch_id(resource)
-
-        filter_entity = filter_item.get("entity_name")
-        filter_pool = filter_item.get("pool_name")
-        filter_switches = filter_item.get("switches") or []
-
-        if filter_entity and not self._entity_names_match(resource_entity, filter_entity):
-            self.log.debug(
-                "manage_gathered: skipping resource id='%s', entity_name mismatch: resource='%s' vs filter='%s'",
-                resource_id,
-                resource_entity,
-                filter_entity,
-            )
-            return False
-
-        if filter_pool and resource_pool != filter_pool:
-            self.log.debug(
-                "manage_gathered: skipping resource id='%s', pool_name mismatch: resource='%s' vs filter='%s'",
-                resource_id,
-                resource_pool,
-                filter_pool,
-            )
-            return False
-
-        if filter_switches and resource_switch_id not in filter_switches:
-            self.log.debug(
-                "manage_gathered: skipping resource id='%s', switchId not in filter: resource_switch='%s', filter_switches=%s",
-                resource_id,
-                resource_switch_id,
-                filter_switches,
-            )
-            return False
-
-        self.log.debug(
-            "manage_gathered: resource id='%s' matched filter (entity_name='%s', pool_name='%s', switch_id='%s')",
-            resource_id,
-            resource_entity,
-            resource_pool,
-            resource_switch_id,
-        )
-        return True
-
-    def _apply_gathered_filters(self):
-        """Apply gathered config filters to cached resources and return translated results."""
-        seen_ids = set()
-        results = []
-
-        for filter_item in self.config:
-            if not self._filter_has_active_criteria(filter_item):
-                self.log.debug(
-                    "manage_gathered: skipping empty filter item (entity_name=%s, pool_name=%s, switches=%s)",
-                    filter_item.get("entity_name"),
-                    filter_item.get("pool_name"),
-                    filter_item.get("switches") or [],
-                )
-                continue
-
-            self.log.debug(
-                "manage_gathered: applying filter: entity_name=%s, pool_name=%s, switches=%s",
-                filter_item.get("entity_name"),
-                filter_item.get("pool_name"),
-                filter_item.get("switches") or [],
-            )
-
-            for resource in self._all_resources:
-                resource_id = self._get_resource_id(resource)
-                if resource_id is not None and resource_id in seen_ids:
-                    self.log.debug(
-                        "manage_gathered: skipping resource id='%s' already included by an earlier filter",
-                        resource_id,
-                    )
-                    continue
-
-                if not self._resource_matches_filter(resource, filter_item):
-                    continue
-
-                translated = self.translate_gathered_results([resource])
-                if translated:
-                    results.append(translated[0])
-                if resource_id is not None:
-                    seen_ids.add(resource_id)
-
-        return results
 
     def manage_gathered(self):
         """Return resources from the ND fabric, optionally filtered by config criteria.
